Pluggable model support, part 2: Plumbing
- The previously unmodifiable enum class ModelClass is replaced with a
QString identifier
- Models register themselves into ModelManager with a proxy object
(derived from ModelProxy) that acts as a hub for all model-related
access

Still missing: Part 3, GUI integration

+++ src/models/Model.cpp
@@ -1,7 +1,7 @@
 #include "Model.h"
 #include "DCThread.h"
 
-Model::Model(ModelPrototype *parent, int instID, DCThread *DCT) :
+Model::Model(ModelPrototype *parent, size_t instID, DCThread *DCT) :
     parent(parent),
     instID(instID)
 {
